fix(Save): Use async refresh
asyncRefresh() was used until now,
but this method relies on a file-watcher.
As a consequence, the following statement
(reloadFromDisk) happens before the file
can be marked as dirty, which makes the first
call to reloadFromDisk() use the idea cache.

Use refreshWithoutFileWatcher() instead, to
make sure that the file is refreshed before
the next call to reloadFromDisk().